Python code cleanup and adding helper scripts to simplify building/deploying Python packages
