+++ .gitignore
@@ -5,5 +5,6 @@ target
 *.pyc
 *.pyo
 *.egg-info
-/build
-*/dist
+/build/
+*/dist/
+*/.tox/
