+++ webapp-python/src/reinvent_sid345/encrypt_decrypt.py
@@ -1,4 +1,19 @@
-""""""
+# Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+# http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+"""Helper class to handle encryption.
+
+This is the only module that you need to modify in the SID345 workshop.
+"""
 import base64
 import json
 import time
@@ -6,10 +21,10 @@ import time
 import aws_encryption_sdk
 
 class EncryptDecrypt(object):
-    """"""
+    """Encrypt and decrypt data."""
 
     def __init__(self, key_id):
-        """"""
+        """Set up materials manager and static values."""
         self._message_type = 'message_type'
         self._type_order_inquiry = 'order inquiry'
         self._timestamp = 'rough timestamp'
@@ -23,7 +38,12 @@ class EncryptDecrypt(object):
         )
 
     def encrypt(self, data):
-        """"""
+        """Encrypt data.
+
+        :param data: JSON-encodeable data to encrypt
+        :returns: Base64-encoded, encrypted data
+        :rtype: str
+        """
         encryption_context = {
             self._message_type: self._type_order_inquiry,
             self._timestamp: str(int(time.time() / 3600.0))
@@ -36,7 +56,11 @@ class EncryptDecrypt(object):
         return base64.b64encode(ciphertext).decode('utf-8')
 
     def decrypt(self, data):
-        """"""
+        """Decrypt data.
+
+        :param bytes data: Base64-encoded, encrypted data
+        :returns: JSON-decoded, decrypted data
+        """
         ciphertext = base64.b64decode(data)
         plaintext, header = aws_encryption_sdk.decrypt(
             source=ciphertext,
