+++ webapp-python/src/reinvent_sid345/handler.py
@@ -1,4 +1,16 @@
-""""""
+# Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+# http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+"""Main Lambda handler for SID345."""
 import base64
 from datetime import datetime, timezone
 from itertools import zip_longest
@@ -9,8 +21,8 @@ import traceback
 
 import boto3
 
-from reinvent_sid345_lambda.encrypt_decrypt import EncryptDecrypt
-from reinvent_sid345_lambda.responses import bad_request, json_error, json_response
+from reinvent_sid345.encrypt_decrypt import EncryptDecrypt
+from reinvent_sid345.responses import bad_request, json_error, json_response
 
 SQS_QUEUE_VAR = 'queue_url'
 KMS_CMK_VAR = 'kms_key_id'
@@ -23,6 +35,7 @@ _is_setup = False
 
 
 def _setup():
+    """Create resources once on Lambda cold start."""
     global _is_setup
     if _is_setup:
         return
@@ -37,33 +50,33 @@ def _setup():
     _is_setup = True
 
 
-def sid345_handler(event, context):
-    """"""
+def lambda_handler(event, context):
+    """Lambda entry point."""
     _LOGGER.debug('REQUEST:')
     _LOGGER.debug(event)
     _setup()
-    if event['httpMethod']  != 'POST':
+    if event['httpMethod'] != 'POST':
         _LOGGER.debug('Bad request method: {}'.format(event['httpMethod']))
         return bad_request()
 
-    response = _handle_post(event)
+    response = handle_post(event)
     _LOGGER.debug('RESPONSE:')
     _LOGGER.debug(response)
     return response
 
 
-def _send(request):
-    """"""
+def send(request):
+    """Handler for send requests."""
     data = request['data']
     ciphertext = _encrypt_decrypt.encrypt(data)
     _sqs_queue.send_message(MessageBody=ciphertext)
     return json_response(200, {'status': 'ok', 'kmsCallCount': 0})
 
 
-def _receive(request):
-    """"""
+def receive(request):
+    """Handler for receive requests."""
     raw_messages = []
-    for _round  in range(MIN_ROUNDS):
+    for _round in range(MIN_ROUNDS):
         messages = _sqs_queue.receive_messages(
             AttributeNames=('SentTimestamp',),
             MaxNumberOfMessages=10,
@@ -76,7 +89,6 @@ def _receive(request):
             # Delete them once we have them?
             raw_messages.append(message)
 
-
     parsed_messages = [_parse_message(message) for message in raw_messages]
     response = {
         'status': 'ok',
@@ -87,13 +99,13 @@ def _receive(request):
 
 
 def _parse_message(message):
-    """"""
+    """Parse and decrypt a SQS message into the expected JSON structure."""
     _LOGGER.debug('MESSAGE:')
     _LOGGER.debug(message)
     _LOGGER.debug('MESSAGE ATTRIBUTES:')
     _LOGGER.debug(message.attributes)
     info = {
-        'timestamp': _iso_format_datetime_from_sent_timestampt(message.attributes['SentTimestamp']),
+        'timestamp': _iso_format_datetime_from_sent_timestamp(message.attributes['SentTimestamp']),
         'messageID': message.message_id,
         'ciphertext': '\n'.join(_data_to_lines(message.body)),
         'decryptInfo': ''
@@ -106,14 +118,14 @@ def _parse_message(message):
     return info
 
 
-def _iso_format_datetime_from_sent_timestampt(sent_timestamp):
-    """"""
+def _iso_format_datetime_from_sent_timestamp(sent_timestamp):
+    """Format a SentTimestamp value from SQS as a ISO8601 datetime."""
     epoch_seconds = int(sent_timestamp) / 1000.0
     return datetime.utcfromtimestamp(epoch_seconds).astimezone(timezone.utc).isoformat()
 
 
 def _data_to_lines(data):
-    """"""
+    """Convert a string of data into 80-character-long lines."""
     return map(
         ''.join,
         zip_longest(
@@ -122,13 +134,15 @@ def _data_to_lines(data):
         )
     )
 
+
 _ACTIONS = {
-    'send': _send,
-    'recv': _receive
+    'send': send,
+    'recv': receive
 }
 
-def _handle_post(event):
-    """"""
+
+def handle_post(event):
+    """Process a POST request."""
     request_body = event['body']
     if event['isBase64Encoded']:
         request_body = base64.b64decode(request_body)
